+++ lib/engineyard-serverside/deploy.rb
@@ -3,6 +3,7 @@ require 'base64'
 require 'fileutils'
 require 'json'
 require 'engineyard-serverside/rails_asset_support'
+require 'ey_instance_api_client'
 
 module EY
   module Serverside
@@ -29,6 +30,7 @@ module EY
         with_failed_release_cleanup do
           create_revision_file
           run_with_callbacks(:bundle)
+          fetch_configs
           symlink_configs
           conditionally_enable_maintenance_page
           run_with_callbacks(:migrate)
@@ -259,6 +261,24 @@ WRAP
         run create_revision_file_command
       end
 
+      def services_fetcher
+        EY::InstanceAPIClient::Services.new
+      end
+
+      def fetch_configs
+        info "~> Fetching configuration resources."
+        services_data = services_fetcher.get(config.app)
+        File.open("#{c.shared_path}/config/ey_services_config_deploy.yml", "w") do |f|
+          YAML.dump(services_data, f)
+        end
+      rescue StandardError => e
+        warning <<-WARNING
+External services configuration not updated. Using previous version.
+Deploy again if your services configuration appears incomplete or out of date.
+#{e}
+        WARNING
+      end
+
       def symlink_configs(release_to_link=c.release_path)
         info "~> Preparing shared resources for release."
         symlink_tasks(release_to_link).each do |what, cmd|
@@ -320,7 +340,7 @@ WRAP
 
       def base_callback_command_for(what)
         [serverside_bin, 'hook', what.to_s,
-          '--app', config.app.to_s,
+          '--app', config.app,
           '--release-path', config.release_path.to_s,
           '--framework-env', c.environment.to_s,
         ].compact