+++ spec/basic_deploy_spec.rb
@@ -4,7 +4,7 @@ describe "Deploying an application without Bundler" do
   before(:all) do
     $DISABLE_GEMFILE = true # Don't generate Gemfile/Gemfile.lock
     $DISABLE_LOCKFILE = true
-    @deploy_dir = File.join(Dir.tmpdir, "serverside-deploy-#{Time.now.to_i}-#{$$}")
+    @deploy_dir = Pathname.new(Dir.tmpdir).join("serverside-deploy-#{Time.now.to_i}-#{$$}")
 
     # set up EY::Serverside::Server like we're on a solo
     EY::Serverside::Server.reset
@@ -13,7 +13,7 @@ describe "Deploying an application without Bundler" do
     # run a deploy
     config = EY::Serverside::Deploy::Configuration.new({
         "strategy"      => "IntegrationSpec",
-        "deploy_to"     => @deploy_dir,
+        "deploy_to"     => @deploy_dir.to_s,
         "group"         => `id -gn`.strip,
         "stack"         => 'nginx_passenger',
         "migrate"       => nil,
@@ -26,24 +26,39 @@ describe "Deploying an application without Bundler" do
     @deployer.deploy
   end
 
+  def exist
+    be_exist
+  end
+
   it "creates a REVISION file" do
-    File.exist?(File.join(@deploy_dir, 'current', 'REVISION')).should be_true
+    @deploy_dir.join('current', 'REVISION').should exist
   end
 
   it "restarts the app servers" do
-    File.exist?(File.join(@deploy_dir, 'current', 'restart')).should be_true
+    @deploy_dir.join('current', 'restart').should exist
   end
 
   it "runs all the hooks" do
-    File.exist?(File.join(@deploy_dir, 'current', 'before_bundle.ran' )).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'after_bundle.ran'  )).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'before_migrate.ran')).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'after_migrate.ran' )).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'before_compile_assets.ran')).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'after_compile_assets.ran' )).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'before_symlink.ran')).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'after_symlink.ran' )).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'before_restart.ran')).should be_true
-    File.exist?(File.join(@deploy_dir, 'current', 'after_restart.ran' )).should be_true
+    @deploy_dir.join('current', 'before_bundle.ran' ).should exist
+    @deploy_dir.join('current', 'after_bundle.ran'  ).should exist
+    @deploy_dir.join('current', 'before_migrate.ran').should exist
+    @deploy_dir.join('current', 'after_migrate.ran' ).should exist
+    @deploy_dir.join('current', 'before_compile_assets.ran').should exist
+    @deploy_dir.join('current', 'after_compile_assets.ran' ).should exist
+    @deploy_dir.join('current', 'before_symlink.ran').should exist
+    @deploy_dir.join('current', 'after_symlink.ran' ).should exist
+    @deploy_dir.join('current', 'before_restart.ran').should exist
+    @deploy_dir.join('current', 'after_restart.ran' ).should exist
+  end
+
+  it "creates and symlinks ey_services_config_deploy.yml" do
+    shared_services_file    = @deploy_dir.join('shared',  'config', 'ey_services_config_deploy.yml')
+    symlinked_services_file = @deploy_dir.join('current', 'config', 'ey_services_config_deploy.yml')
+
+    shared_services_file.should exist
+    shared_services_file.should_not be_symlink
+
+    symlinked_services_file.should exist
+    symlinked_services_file.should be_symlink
   end
 end