+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 
+  * Add ey_services_config_deploy.yml file through EY::InstanceAPI during deploy.
   * Bump default bundler version to 1.0.21
 
 ## v1.5.10 (2011-11-01)